refactor external txs block
