+++ ui/tx/details/TxDetails.tsx
@@ -2,7 +2,6 @@ import {
   Box,
   GridItem,
   Text,
-  Spinner,
   Flex,
   chakra,
   VStack,
@@ -18,8 +17,6 @@ import { ZKSYNC_L2_TX_BATCH_STATUSES } from 'types/api/zkSyncL2';
 import { route } from 'nextjs-routes';
 
 import config from 'configs/app';
-import useApiQuery from 'lib/api/useApiQuery';
-import useIsMobile from 'lib/hooks/useIsMobile';
 import getNetworkValidatorTitle from 'lib/networks/getNetworkValidatorTitle';
 import * as arbitrum from 'lib/rollups/arbitrum';
 import { formatZkEvmTxStatus, formatZkSyncL2TxnBatchStatus, layerLabels } from 'lib/rollups/utils';
@@ -60,7 +57,6 @@ import TxDetailsTokenTransfers from 'ui/tx/details/TxDetailsTokenTransfers';
 import TxDetailsWithdrawalStatusOptimistic from 'ui/tx/details/TxDetailsWithdrawalStatusOptimistic';
 import TxRevertReason from 'ui/tx/details/TxRevertReason';
 import TxAllowedPeekers from 'ui/tx/TxAllowedPeekers';
-import TxExternalTxs from 'ui/tx/TxExternalTxs';
 import TxSocketAlert from 'ui/tx/TxSocketAlert';
 import ZkSyncL2TxnBatchHashesInfo from 'ui/txnBatches/zkSyncL2/ZkSyncL2TxnBatchHashesInfo';
 
@@ -72,6 +68,7 @@ import TxDetailsSetMaxGasLimit from './TxDetailsSetMaxGasLimit';
 import TxDetailsTacOperation from './TxDetailsTacOperation';
 import TxDetailsTxFee from './TxDetailsTxFee';
 import TxDetailsWithdrawalStatusArbitrum from './TxDetailsWithdrawalStatusArbitrum';
+import TxHash from './TxHash';
 import TxInfoScrollFees from './TxInfoScrollFees';
 
 interface Props {
@@ -81,23 +78,10 @@ interface Props {
   noTxActions?: boolean;
 }
 
-const externalTxFeature = config.features.externalTxs;
 const rollupFeature = config.features.rollup;
 
 const TxDetails = ({ data, isLoading, socketStatus, noTxActions }: Props) => {
   const [ isExpanded, setIsExpanded ] = React.useState(false);
-
-  const isMobile = useIsMobile();
-
-  const externalTxsQuery = useApiQuery('general:tx_external_transactions', {
-    pathParams: {
-      hash: data?.hash,
-    },
-    queryOptions: {
-      enabled: externalTxFeature.isEnabled,
-      placeholderData: [ '1', '2', '3' ],
-    },
-  });
 
   const handleCutLinkClick = React.useCallback(() => {
     setIsExpanded((flag) => !flag);
@@ -131,6 +115,7 @@ const TxDetails = ({ data, isLoading, socketStatus, noTxActions }: Props) => {
       </chakra.span>
     </Tooltip>
   ) : null;
+
   const executionFailedBadge = toAddress?.is_contract && Boolean(data.status) && data.result !== 'success' ? (
     <Tooltip content="Error occurred during contract execution">
       <chakra.span display="inline-flex" ml={ 2 } mr={ 1 }>
@@ -166,33 +151,7 @@ const TxDetails = ({ data, isLoading, socketStatus, noTxActions }: Props) => {
 
       { config.features.crossChainTxs.isEnabled && <TxDetailsCrossChainMessages hash={ data.hash } isLoading={ isLoading }/> }
 
-      <DetailedInfo.ItemLabel
-        hint="Unique character string (TxID) assigned to every verified transaction"
-        isLoading={ isLoading }
-      >
-        Transaction hash
-      </DetailedInfo.ItemLabel>
-      <DetailedInfo.ItemValue multiRow={ config.features.externalTxs.isEnabled && externalTxsQuery.data && externalTxsQuery.data.length > 0 }>
-        <Flex flexWrap="nowrap" alignItems="center" overflow="hidden">
-          { data.status === null && <Spinner mr={ 2 } size="sm" flexShrink={ 0 }/> }
-          <Skeleton loading={ isLoading } overflow="hidden">
-            <HashStringShortenDynamic hash={ data.hash }/>
-          </Skeleton>
-          <CopyToClipboard text={ data.hash } isLoading={ isLoading }/>
-          { config.features.metasuites.isEnabled && (
-            <>
-              <TextSeparator flexShrink={ 0 } display="none" id="meta-suites__tx-explorer-separator"/>
-              <Box display="none" flexShrink={ 0 } id="meta-suites__tx-explorer-link"/>
-            </>
-          ) }
-        </Flex>
-        { config.features.externalTxs.isEnabled && externalTxsQuery.data && externalTxsQuery.data.length > 0 && (
-          <Skeleton loading={ isLoading || externalTxsQuery.isPlaceholderData } display={{ base: 'block', lg: 'inline-flex' }} alignItems="center">
-            { !isMobile && <TextSeparator flexShrink={ 0 }/> }
-            <TxExternalTxs data={ externalTxsQuery.data }/>
-          </Skeleton>
-        ) }
-      </DetailedInfo.ItemValue>
+      <TxHash hash={ data.hash } isLoading={ isLoading } status={ data.status }/>
 
       <DetailedInfo.ItemLabel
         hint="Current transaction state: Success, Failed (Error), or Pending (In Process)"
